Added file StockSpan.java
The stock span problem is a financial problem where we have a series of n daily price quotes for a stock and we need to calculate the span of stock’s price for each n days.
The problem can be solved using the concept of "Nearest Smaller to Left" problem. Only difference is that we need to track the INDICES of the smaller element instead of the actual element. So, we need Stack to track the smaller element and an ArrayList to store the result.