Create Queue
Queue is a Data Structure that works on the concept of FIFO (First In First Out).
This class demonstrate simple functionalities such as - enqueue (insertion of data), dequeue (deletion of data) and finds if the queue is empty or not.
Internally, array is used so initially the user has to input the size also. Therefore overflow and underflow can also occur.