Queue
Queue is a simple data structure that works on the principle of FIFO (First In First Out).
This class demonstrate very basic functions of queue i.e. enqueue (insertion), dequeue (deletion), isEmpty (checks if queue if empty or not.)
Internally it uses array so the user has to give size or capacity as input. So overflow and underflow can also occur.