Circular Queue
Circular Queue is similar to regular Queue except that the last position is connected back to the first position to make a circle.
Implementation is done same as the regular queue. 
In the above code, in the main method the input is hard coded and one should expect the result in accordance to the hard coded data.